+++ 08_upsert_observations.sql
@@ -7,6 +7,7 @@ with id_synthese as common field.
  */
 BEGIN;
 
+
 DROP TRIGGER IF EXISTS tri_c_upsert_vn_observation_to_geonature ON
     src_vn_json.observations_json;
 
@@ -789,5 +790,26 @@ CREATE TRIGGER tri_c_delete_vn_observation_from_geonature
     AFTER DELETE ON src_vn_json.observations_json
     FOR EACH ROW
     EXECUTE PROCEDURE src_lpodatas.fct_tri_c_delete_vn_observation_from_geonature ();
+
+
+CREATE OR REPLACE FUNCTION src_lpodatas.fct_c_update_user_observations(_observer_uid text) RETURNS void
+    LANGUAGE plpgsql
+AS
+$$
+DECLARE
+    row_count INT;
+BEGIN
+    RAISE DEBUG '<fct_c_update_user_observations> Update observations for user %', _observer_uid;
+    WITH "rows" AS (
+        UPDATE src_vn_json.observations_json
+            SET item = item
+            WHERE observations_json.item #>> '{observers,0,@uid}' = _observer_uid
+            RETURNING 1)
+    SELECT count(*) into row_count
+    FROM "rows";
+    RAISE DEBUG '<fct_c_update_user_observations> % Update observations for user %', row_count, _observer_uid;
+END;
+$$;
+
 
 COMMIT;